+++ test/test-node-cdp.js
@@ -142,8 +142,7 @@ const generate = async () => {
     console.log('generate v8-node cdp coverage reports ...');
 
     const coverageResults = await new CoverageReport(coverageOptions).generate();
-    console.log('reportPath', EC.magenta(coverageResults.reportPath));
-    console.log('v8-node cdp coverage generated', Object.keys(coverageResults.summary).map((k) => [k, coverageResults.summary[k].pct]));
+    console.log('v8-node cdp coverage reportPath', EC.magenta(coverageResults.reportPath));
 
     ws.close();
 };