+++ test/test-node-ins.js
@@ -111,8 +111,7 @@ const generate = async () => {
     console.log('generate v8-node coverage reports ...');
 
     const coverageResults = await new CoverageReport(coverageOptions).generate();
-    console.log('reportPath', EC.magenta(coverageResults.reportPath));
-    console.log('v8-node coverage generated', Object.keys(coverageResults.summary).map((k) => [k, coverageResults.summary[k].pct]));
+    console.log('v8-node coverage reportPath', EC.magenta(coverageResults.reportPath));
 
 };